backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
httpx
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
